requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.4
django-celery-results==2.5.1
djangorestframework==3.14.0
env==0.1.0
environ==1.0
exceptiongroup==1.1.3
frozenlist==1.4.0
idna==3.4
iniconfig==2.0.0
kombu==5.3.1
magic-filter==1.0.11
multidict==6.0.4
packaging==23.1
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
pytest-factoryboy==2.5.1
factory-boy==3.3.0
faker==18.13.0
